tests: Removed the ignoring of tests
This commit reverts
https://gem5-review.googlesource.com/c/public/gem5/+/23023.

It has proven to be an unpopular piece of functionality which makes it
too easy to silently ignore failing tests. The new policy will be to
remove/comment-out failing tests in the testing source and tag with Jira
issues as to why.

Change-Id: I17d69fc57a9171ce3702e019615390a9aa3da250
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/33139
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jason Lowe-Power <power.jg@gmail.com>
Maintainer: Jason Lowe-Power <power.jg@gmail.com>

Former-commit-id: 7c82d1b8d5db9081306628a9d2024b7aaf3ec63c
Former-commit-id: 0e1ddc31fc36bc7741062ff55611c731da09b751